+++ test/td_dd_web/controllers/grant_request_approval_controller_test.exs
@@ -1,4 +1,4 @@
-defmodule TdDdWeb.ApprovalControllerTest do
+defmodule TdDdWeb.GrantRequestApprovalControllerTest do
   use TdDdWeb.ConnCase
 
   describe "create" do